database::lib.rs into 4 files.
